+++ opentmf-api-clients-hub/src/main/java/org/opentmf/api/client/hub/config/ReactiveTmfHubAutoConfiguration.java
@@ -5,6 +5,7 @@ import static org.opentmf.client.common.util.TokenUtil.TOKEN_SERVICE;
 import static org.opentmf.client.common.util.TokenUtil.WEB_CLIENT;
 
 import lombok.extern.slf4j.Slf4j;
+import org.jspecify.annotations.NonNull;
 import org.opentmf.api.client.common.config.TmfApiClientsConfig;
 import org.opentmf.api.client.common.util.TmfApiClientConstants;
 import org.opentmf.api.client.hub.api.ReactiveTmfHubClient;
@@ -40,7 +41,7 @@ public class ReactiveTmfHubAutoConfiguration {
     return new BeanDefinitionRegistryPostProcessor() {
 
       @Override
-      public void postProcessBeanDefinitionRegistry(@org.jspecify.annotations.NonNull BeanDefinitionRegistry registry)
+      public void postProcessBeanDefinitionRegistry(@NonNull BeanDefinitionRegistry registry)
           throws BeansException {
 
         TmfApiClientsConfig config = Binder.get(env)
@@ -83,7 +84,7 @@ public class ReactiveTmfHubAutoConfiguration {
       }
 
       @Override
-      public void postProcessBeanFactory(@org.jspecify.annotations.NonNull ConfigurableListableBeanFactory beanFactory)
+      public void postProcessBeanFactory(@NonNull ConfigurableListableBeanFactory beanFactory)
           throws BeansException {
         // no-op
       }